interview/model_eval/pyproject.toml
[project]
name = "flyte-production"
version = "0.1.0"
description = "A Flyte production project"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "flytekit",
    "flytekitplugins-deck-standard>=1.16.10",
    "pandas>=2.3.3",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = [
    "isort>=6.1.0",
    "ruff>=0.14.6",
]
